+++ server/db.json
@@ -411,21 +411,10 @@
   ],
   "carts": {
     "cart_id": 1,
-    "items": [
-      {
-        "categoryName": "IT/프로그래밍",
-        "subCategoryName": "프론트엔드",
-        "courseId": 2003,
-        "courseTitle": "TypeScript 마스터클래스",
-        "instructorName": "박강사",
-        "price": 75000,
-        "coverImgUrl": "https://images.unsplash.com/photo-1516116216624-53e697fedbea?auto=format&fit=crop&w=1200&q=80",
-        "addedAt": "2025-12-11T05:55:46.755Z"
-      }
-    ],
+    "items": [],
     "summary": {
-      "totalCount": 1,
-      "totalAmount": 75000
+      "totalCount": 0,
+      "totalAmount": 0
     }
   },
   "reviews": [
