+++ src/features/auth/Action/logout.action.ts
@@ -1,0 +1,10 @@
+"use server";
+
+import { cookies } from "next/headers";
+import { redirect } from "next/navigation";
+
+export const logoutAction = async () => {
+  const cookieStore = await cookies();
+  cookieStore.delete("accessToken");
+  redirect("/courses");
+};
