+++ src/features/mypage/components/MySidebar.tsx
@@ -2,7 +2,7 @@
 
 import Link from "next/link";
 import styles from "./MySidebar.module.css";
-import { logoutAction } from "@/features/auth/actions";
+import { logoutAction } from "@/features/auth/Action/logout.action";
 
 const userProfile = {
   displayName: "윤선",
